[MISC] More 64b fixes but those for doom classic, caught with basic warning 4311: pointer truncation

+++ neo/sys/win32/win_cpu.cpp
@@ -97,7 +97,7 @@ double Sys_ClockTicksPerSecond() {
 			}
 			RegCloseKey( hKey );
 			if ( ret == ERROR_SUCCESS ) {
-				ticks = (double) ((unsigned long)ProcSpeed) * 1000000;
+				ticks = (double) ((uintptr_t)ProcSpeed) * 1000000;
 			}
 		}
 	}
